+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 ## [Unreleased] - 2021-08-28
+### Added
+- add support for extracting issue numbers for changelog summary
+
 ### Changed
 - rework directory structure and update .gitignore
 - initial version